requirements.txt
streamlit>=1.36.0
flask>=3.0.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
PyMuPDF>=1.24.2
langchain>=0.2.7
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
duckduckgo-search>=6.2.11
arxiv>=2.1.0

